Only delete images that aren't being used for #3707

+++ app/models/work.rb
@@ -280,10 +280,19 @@ class Work < ApplicationRecord
   end
 
   def cleanup_images
+    absolute_filenames = pages.map { |page| [page.base_image, page.thumbnail_filename]}.flatten
+    modern_filenames = absolute_filenames.map{|fn| fn.sub(/^.*uploaded/, File.join(Rails.root, "public", "images", "uploaded"))}
+    binding.pry
+    modern_filenames.each do |fn|
+      File.delete(fn) if File.exist?(fn)
+    end
     new_dir_name = File.join(Rails.root, "public", "images", "uploaded", self.id.to_s)
     if Dir.exist?(new_dir_name)
-      Dir.glob(File.join(new_dir_name, "*")){|f| File.delete(f)}
-      Dir.rmdir(new_dir_name)
+      # test to see if the directory is empty
+      if Dir.glob(File.join(new_dir_name, "*")).empty?
+        # if it is, delete it
+        Dir.rmdir(new_dir_name)
+      end
     end
   end
